+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+## [1.8.0] - 2024-04-29
+### Breaking changes
+- Removed (deprecated) networks Goerli, Op Goerli, Arb Goerli, Mumbai, Klaytn and Mantle Testnet (Goerli)
+- Added OP Sepolia, Arb Sepolia, Mantle Sepolia, Polygon Amoy testnet,  Ancient8 mainnet, Ancient8 testnet supported networks
+- Changed testnet bundler rpc urls to the new format
+
 ## [1.7.1] - 2024-04-12
 ### New
 - Added `getSupportedAssets` to get supported tokens